% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/demo.R
\name{demo.pollen.run}
\alias{SEHmodel-demo}
\alias{demo.pollen.run}
\title{SEHmodel package pollen demo}
\usage{
demo.pollen.run(nb_fields=200,max_size=5000,raster_size=2^10,nb_ind=100)
}
\arguments{
\item{nb_fields}{number of fields in the landscape (default 200)}

\item{max_size}{landscape size (in meter) (default 5000)}

\item{raster_size}{raster size (default 2^10)}

\item{nb_ind}{number of individuals to simulate (default 100)}
}
\description{
Demo of SEHmodel package on Bt maize pollen (genetically modified crop) on non target butterfly species larvae.
}
\keyword{demo}
\keyword{demo.pollen.run}

