% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecoToxic.R
\docType{methods}
\name{ecoToxic}
\alias{ecoToxic}
\alias{ecoToxic,Landscape,Individuals-method}
\alias{ecotoxicological-method}
\title{Method ecoToxicological}
\usage{
ecoToxic(objectL, objectI, ...)

\S4method{ecoToxic}{Landscape,Individuals}(objectL, objectI, objectT,
  mintime = 1, maxtime, kin = 0.25, kout = 0.5, deltat = 0.1)
}
\arguments{
\item{objectL}{A Landscape object}

\item{objectI}{An Individuals object}

\item{...}{other parameters}

\item{objectT}{A 3D array of Toxic Dispersion over time [t,x,y], first indice is time}

\item{mintime}{Time to start simulation (default = 1)}

\item{maxtime}{Time to end simulation}

\item{kin}{\% of toxic ingestion}

\item{kout}{\% of toxic elimination}

\item{deltat}{\% of a time step}
}
\value{
An Individuals object
}
\description{
Generic method on Landscape and Individuals objects for ecotoxicological.

Internal concentration of contaminants within individuals
}

