% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individuals-Methods.R
\docType{methods}
\name{[,Individuals,numeric,numeric,logical-method}
\alias{[,Individuals,numeric,ANY,ANY-method}
\alias{[,Individuals,numeric,numeric,logical-method}
\title{Get an individual infomation}
\usage{
\S4method{[}{Individuals,numeric,numeric,logical}(x, i, j, ..., drop = T)
}
\arguments{
\item{x}{An Individuals object}

\item{i}{individual index}

\item{j}{time of information}

\item{...}{further arguments passed to or from others methods.}

\item{drop}{logical value (default = TRUE)}
}
\value{
a data.frame
}
\description{
Get an Individual information
}

