% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Individuals-Methods.R
\docType{methods}
\name{getIndividualsLife}
\alias{getIndividualsLife}
\alias{getIndividualsLife,Individuals-method}
\title{Method to get Individuals Life info}
\usage{
getIndividualsLife(object, ...)

\S4method{getIndividualsLife}{Individuals}(object)
}
\arguments{
\item{object}{An Inidividuals object}

\item{...}{others parameters}
}
\value{
a matrix indexed by individual ID in row and by time in columns.
}
\description{
Method to get Individuals Life info

Get individuals toxic concentration over the simulation time.
If intern concentration overtakes the toxic threshold value is "-2", that means the toxic kill the individual
otherwise value is "-1" means the individual is dead in natural way. The value "0" means that the individual is not alive yet.
}

