% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Individuals-Methods.R
\docType{methods}
\name{plot,Individuals,ANY-method}
\alias{plot,Individuals,ANY-method}
\alias{plot,Individuals,num-method}
\alias{plot,Individuals,numeric-method}
\title{Method plot}
\usage{
\S4method{plot}{Individuals,ANY}(x, y, add = F, ..., plot.legend = TRUE)

\S4method{plot}{Individuals,numeric}(x, y, add = F, ..., plot.legend = TRUE)
}
\arguments{
\item{x}{An Individuals object}

\item{y}{time of the simulation to display individuals}

\item{add}{if True will plot hover an already plot image (default False)}

\item{...}{further graphical parameters (\code{par})}

\item{plot.legend}{plot legend (default TRUE)}
}
\description{
Will plot individuals spatial positions.

Will plot individuals positions and state at a time of the simulation.
}
\details{
"Red" cross means that the individual is dead by toxic exposition.
"White" cross means that the individual is dead by natural death.
"Green" to "Red" points give the gradiant of toxic concentration before the threshold.
}

