% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Individuals-Methods.R
\docType{methods}
\name{print,Individuals-method}
\alias{print,Individuals-method}
\title{print}
\usage{
\S4method{print}{Individuals}(x, ...)
}
\arguments{
\item{x}{An Individuals object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print
}

