% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Individuals-Methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,Individuals-method}
\title{Show part of Individuals Info}
\usage{
\S4method{show}{Individuals}(object)
}
\arguments{
\item{object}{An Individuals object}
}
\description{
Show part of Individuals Info
}

