% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class-Landscape.R
\docType{class}
\name{Landscape-class}
\alias{Landscape-class}
\title{Class Landscape}
\description{
Class \code{Landscape} defines a landscape
}
\details{
Objects can be created by calling of the allocator new("Landscape", ...), or (preferred) by calling to the function \code{simulateLandscape} or \code{loadLandscpae}.
}
\section{Slots}{

\describe{
\item{\code{thelandscape}}{A SpatialPolygonsDataFrame}

\item{\code{xmin}}{Left x coordinates (in meters)}

\item{\code{xmax}}{Right x coordinates (in meters)}

\item{\code{ymin}}{Bottom y coordinates (in meters)}

\item{\code{ymax}}{Top y coordinates (in meters)}

\item{\code{n}}{Number of fields in the landscape}
}}
\seealso{
\code{simulateLandscpae} , \code{loadLandscape} , \code{loadLandscapeSIG}
}

