% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{simulateInitialPartition}
\alias{simulateInitialPartition}
\title{function simulateInitialPartition

Create an object \code{Landscape}. Simulate a landscape with neutral and source fields.}
\usage{
simulateInitialPartition(n = 500, prop = 0.4, range = 10, xmin = 0,
  xmax = 5000, ymin = 0, ymax = 5000)
}
\arguments{
\item{n}{Numeric, numbers of fields}

\item{prop}{Numeric [0,1] toxic fields proportion}

\item{range}{truc}

\item{xmin}{x left coordinates}

\item{xmax}{x right coordinates}

\item{ymin}{y top coordinates}

\item{ymax}{y bottom coordinates}
}
\value{
An S4 \code{Landscape} object with n fields, prop pourcentage of toxic fields of size (xmin,xmax) (ymin,ymax)
}
\description{
function simulateInitialPartition

Create an object \code{Landscape}. Simulate a landscape with neutral and source fields.
}
\examples{
simulateInitialPartition(n=500,prop=0.4,range=10,xmin=0,xmax=5000,ymin=0,ymax=5000)
}

