% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\docType{methods}
\name{getSPSources}
\alias{Landscape,getSPSources-method}
\alias{getSPSources}
\alias{getSPSources,Landscape-method}
\title{Method getSPSources}
\usage{
getSPSources(object, ...)

\S4method{getSPSources}{Landscape}(object)
}
\arguments{
\item{object}{A Landscape object}

\item{...}{options}
}
\value{
a SpatialPolygonDataFrame object
}
\description{
Method getSPSources

Get all polygons of a landscape object identify as sources
}

