% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{loadLandscape}
\alias{loadLandscape}
\title{Wrapper function loadLandscape}
\usage{
loadLandscape(sp, data)
}
\arguments{
\item{sp}{a SpatialPolygons object}

\item{data}{a data.frame containing fields (polygons) informations. Row num as fields ID, cols names as sources | neutral | receptors (1 if fields are of this types 0 otherwise)}
}
\description{
Wrapper to create a Landscape object using SpatialPolygons and dataframe. The SpatialPolygons object and the data.frame have to contain the same number of polygons and row (row ID is polygons ID).
}

