% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{loadLandscapeSIG}
\alias{loadLandscapeSIG}
\title{Create a Landscape object from SIG shapefile file}
\usage{
loadLandscapeSIG(dsn, layer)
}
\arguments{
\item{dsn}{folder path to the source files}

\item{layer}{file name without extension}
}
\description{
Create a Landscape object from SIG shapefile. Shapefile had to contain a SpatialPolygonsDataFrame. Data in the data frame containing fields (polygons) informations. Row num as fields ID, cols names as sources | neutral | receptors (1 if fields are of this types 0 otherwise).
}

