% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\docType{methods}
\name{print,Landscape-method}
\alias{print,Landscape-method}
\title{print Landscape info
print}
\usage{
\S4method{print}{Landscape}(x, ...)
}
\arguments{
\item{x}{A Landscape object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print Landscape info
print
}

