% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\docType{methods}
\name{show,Landscape-method}
\alias{show,Landscape-method}
\title{show Landscape Info}
\usage{
\S4method{show}{Landscape}(object)
}
\arguments{
\item{object}{A Landscape class}
}
\description{
show Landscape Info
}

