% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\docType{methods}
\name{simulateThickMargins}
\alias{simulateThickMargins}
\alias{simulateThickMargins

Simulate Border in a landscape.
Border width use zero-inflated distribution}
\alias{simulateThickMargins,Landscape-method}
\title{Method simulateThickMargins}
\usage{
simulateThickMargins(objectL, ...)

\S4method{simulateThickMargins}{Landscape}(objectL, border_size = 200,
  prob = runif(1, 0.1, 0.9), mean_thickness = runif(1, 2, 20),
  v_thickness = 50)
}
\arguments{
\item{objectL}{A Landscape object}

\item{...}{other parameters}

\item{border_size}{A numeric, bbox margin}

\item{prob}{Probability to inflated a border}

\item{mean_thickness}{Border width expectation}

\item{v_thickness}{Border width variance}
}
\value{
a Landscape object with border as receptor
}
\description{
Method simulateThickMargins
}

