% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Individuals-Methods.R
\name{loadIndividuals}
\alias{loadIndividuals}
\title{Wrapper function loadIndividuals}
\usage{
loadIndividuals(objectL, sp, data, mintime, maxtime)
}
\arguments{
\item{objectL}{a Landscape object}

\item{sp}{a SpatialPoint object (individuals coordinates)}

\item{data}{a data.frame containing individuals attributs. Row num as individuals ID, cols names as dob (date of birth) | life_duration | toxic_threshold}

\item{mintime}{Start simulation time}

\item{maxtime}{End simulation time}
}
\value{
an Individuals object
}
\description{
Wrapper to create an Individuals object using SpatialPoints and dataframe.

The SpatialPoints object and the data.frame have to contain the same number of coordinates and row.
}

