% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individuals-Methods.R
\name{simulateIndividuals}
\alias{simulateIndividuals}
\title{Wrapper function SimulateIndividuals}
\usage{
simulateIndividuals(objectL, n = 200, mintime = 1, maxtime, dob,
  life_duration, toxic_threshold)
}
\arguments{
\item{objectL}{A Landscape object}

\item{n}{Number of individuals to simulate}

\item{mintime}{Start simulation time}

\item{maxtime}{End simulation time}

\item{dob}{A vector for the Date Of Birth of each individual}

\item{life_duration}{A vector for the life duration of each individual}

\item{toxic_threshold}{A vector for the intern toxic threshold value leading to death for each individual}
}
\value{
A S4 \code{Individuals} object
}
\description{
This function simulates individuals as an Individuals object.

Will simulate \code{nb} individuals in receptors fields of a \code{landscape}.
}
\details{
The Individuals object output includes for each individual the coordinates, the date of birth, the life duration and the intern toxic concentration.
}
\seealso{
\link{loadIndividuals}
}

