% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Individuals-Methods.R
\docType{methods}
\name{[,Individuals,numeric,numeric,logical-method}
\alias{[,Individuals,numeric,numeric,logical-method}
\title{Get one individual infos}
\usage{
\S4method{[}{Individuals,numeric,numeric,logical}(x, i, j, ..., drop = T)
}
\arguments{
\item{x}{An Individuals object}

\item{i}{individual index}

\item{j}{time of informations}

\item{...}{further arguments passed to or from others methods.}

\item{drop}{logical value (default = TRUE)}
}
\description{
Get one individual infos
}

