% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Individuals-Methods.R
\docType{methods}
\name{[, Individuals}
\alias{[, Individuals}
\alias{[,Individuals,numeric,ANY,ANY-method}
\title{Select subset of individuals data}
\usage{
\S4method{[}{Individuals,numeric,ANY,ANY}(x, i, j, ..., drop = T)
}
\arguments{
\item{x}{An Individuals object}

\item{i}{individual index}

\item{j}{time of informations}

\item{...}{further arguments passed to or from others methods.}

\item{drop}{logical value (default = TRUE)}
}
\value{
a data.frame
}
\description{
Select subset of individuals data
}

