% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individuals-Methods.R
\docType{methods}
\name{Individuals plot}
\alias{Individuals plot}
\alias{plot,Individuals,ANY-method}
\alias{plot,Individuals,num-method}
\alias{plot,Individuals,numeric-method}
\title{Plot method for Individuals}
\usage{
\S4method{plot}{Individuals,ANY}(x, y, add = F, ..., plot.legend = TRUE)

\S4method{plot}{Individuals,numeric}(x, y, add = F, ..., plot.legend = TRUE)
}
\arguments{
\item{x}{An Individuals object}

\item{y}{time of the simulation to display individuals}

\item{add}{if True the new plot will overlap an already plot image (default False)}

\item{...}{further graphical parameters (\code{par})}

\item{plot.legend}{plot legend (default TRUE)}
}
\description{
Will plot individuals spatial positions.

Will plot individuals positions and state at a time of the simulation.
}
\details{
"Red" cross means that the individual is dead becouse of toxic exposition.
"Green" cross means that the individual is dead by natural death.
"Green" to "Red" points give the gradient of toxic concentration before the threshold.
}

