% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Individuals-Methods.R
\docType{methods}
\name{plot,Landscape,Individuals-method}
\alias{plot,Landscape,Individuals-method}
\title{plot Landscape & Individuals method}
\usage{
\S4method{plot}{Landscape,Individuals}(x, y, time = -1, objectT = NULL,
  numind = -1, add = F, ..., plot.legend = TRUE)
}
\arguments{
\item{x}{A Landscape object}

\item{y}{An Individuals object}

\item{time}{Time selection (default = -1) for Toxic intensity}

\item{objectT}{ToxicIntensityRaster, 3D array (result of \link{toxicIntensity}, default NULL)}

\item{numind}{an individual ID}

\item{add}{Boolean to draw hover an another plot}

\item{...}{default plot parameters (par)}

\item{plot.legend}{plot legend (default TRUE)}
}
\description{
Plots using Landscape & Individuals objects.
}
\details{
If objectT and numind are informed, the function will draw an individual ecoToxical plot.

 If objectT and time are specified, the function will draw the landscape, the individuals states and the toxic intensity at this time of the simulation.
 
 By default this function will draw the landscape and the individuals positions.
}

