% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\docType{methods}
\name{plot,Landscape,ANY-method}
\alias{plot,Landscape,ANY-method}
\title{Landscape plot Method}
\usage{
\S4method{plot}{Landscape,ANY}(x, y, add = F, objectT = NULL, time = -1,
  ..., plot.legend = TRUE)
}
\arguments{
\item{x}{A Landscape object}

\item{y}{ANY}

\item{add}{boolean to add draw hover plot}

\item{objectT}{Toxic intensity matrix (result of \link{toxicIntensity}, default NULL)}

\item{time}{Time selection (default = -1) for Toxic intensity}

\item{...}{default plot par}

\item{plot.legend}{plot legend (default TRUE)}
}
\description{
Landscape plot Method
}

