% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{saveIntoFile}
\alias{saveIntoFile}
\title{Save Particles Dispersion 3D Array to tiff file}
\usage{
saveIntoFile(objectL, objectT, filename = "ParticlesDispersion.tif",
  format = "GTiff", overwrite = T)
}
\arguments{
\item{objectL}{a Landscape object}

\item{objectT}{a 3D array particles dispersion indexed by time}

\item{filename}{output file name}

\item{format}{output format (default=GTiff)}

\item{overwrite}{if True overwrite filename}
}
\value{
a RasterStack object
}
\description{
Save into tiff file particles dispersion 3D array from toxicIntensity.
The output is a RasterStack with a layer per time unit with projection set to CRS="+proj=longlat +datum=WGS84"
}

