% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecoToxic.R
\name{plotEcotoxic}
\alias{plotEcoToxic}
\alias{plotEcotoxic}
\title{Plot internal toxic concentration}
\usage{
plotEcoToxic(objectL, objectI, objectT, numind = 8)
}
\arguments{
\item{objectL}{A Landscape object}

\item{objectI}{An Individuals object}

\item{objectT}{A 3d array of Toxic intensity over the time [t,x,y], (first indice is time) see \code{\link{toxicIntensity}}}

\item{numind}{An individual id}
}
\description{
Plot a time serie of internal toxic concentration for a given individual.
}

