% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicIntensity.R
\name{plotLandscapetoxicIntensity}
\alias{plotLandscapetoxicIntensity}
\title{PlotLandscapetoxicIntensity Method}
\usage{
plotLandscapetoxicIntensity(objectL, objectT, time)
}
\arguments{
\item{objectL}{a Landscape object}

\item{objectT}{a 3D array from toxicIntensity, [time,x,y]}

\item{time}{time to plot the toxic dispersion}
}
\description{
Plot a landscape and the toxic intensity at a time t.
}

