% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Individuals-Methods.R
\name{simulateIndividuals}
\alias{simulateIndividuals}
\title{Wrapper function SimulateIndividuals}
\usage{
simulateIndividuals(objectL, n = 200, mintime = 1, maxtime, dob,
  life_duration, toxic_threshold)
}
\arguments{
\item{objectL}{A Landscape object}

\item{n}{Number of individuals to simulate}

\item{mintime}{Start simulation time}

\item{maxtime}{End simulation time}

\item{dob}{A vector for the Date Of Birth of each individual}

\item{life_duration}{A vector for the life duration of each individual}

\item{toxic_threshold}{A vector for the intern toxic threshold value for each individual before death}
}
\value{
A S4 \code{Individuals} object
}
\description{
Wrapper to create an Individuals object.

Will simulate \code{nb} individuals in receptors fields of a \code{landscape}.
}

